Remove port.h, merge with Platform.h, rework min & max
Removed the port.h header
Merged port.h useful parts with Platform.h
Reworked min and max macros to use std::min and std::max instead
Fixed all uses of min and max that didn't use the same types
Removed Windows.h includes in other places, relying on Platform.h only
Removed HMODULE and GetProcAddress definitions; cross-platform versions
exist
Tidied up interface.h and interface.cpp a bit
Added TODO items

+++ game/server/entities/items/CItemBattery.cpp
@@ -40,7 +40,7 @@ bool CItemBattery::MyTouch( CBasePlayer *pPlayer )
 		char szcharge[ 64 ];
 
 		pPlayer->pev->armorvalue += gSkillData.batteryCapacity;
-		pPlayer->pev->armorvalue = min( pPlayer->pev->armorvalue, MAX_NORMAL_BATTERY );
+		pPlayer->pev->armorvalue = min( pPlayer->pev->armorvalue, static_cast<float>( MAX_NORMAL_BATTERY ) );
 
 		EMIT_SOUND( pPlayer->edict(), CHAN_ITEM, "items/gunpickup2.wav", 1, ATTN_NORM );
